fix corrupted gifs

+++ run.sh
@@ -62,10 +62,12 @@ main() {
   
   # Removing old volumes
   log "Removing old volumes"
-  docker-compose down -v >> "$LOGFILE" 2>&1 || {
-    log "Error removing old volumes. Check $LOGFILE for details."
-    exit 1
-  }
+  if [ "$(docker ps -q)" ]; then
+    docker-compose down -v >> "$LOGFILE" 2>&1 || {
+      log "Error removing old volumes. Check $LOGFILE for details."
+      exit 1
+    }
+  fi
   # Build Docker containers
   log "Building Docker containers"
   docker-compose build --no-cache $BUILD_ARGS >> "$LOGFILE" 2>&1 || {
